fix some suggestion
